refactor: inline nested story components and remove stories/** exclusions
- Inline nested story components from NVD3 BoxPlot, Bubble, Bullet, and
  Compare into their parent *.stories.tsx files
- Rename Screengrid Stories.tsx to follow *.stories.tsx convention
- Remove src/**/stories/** exclusions from 19 plugin tsconfig files
  since nested non-story files no longer exist
- Delete the now-empty stories/stories/ directories

This allows data.ts helper files in story folders to be type-checked
while story files themselves remain excluded via the *.stories.*
pattern.
